extension: refac splitting providers into new files
