+++ src/extension.ts
@@ -1,13 +1,8 @@
 import * as vscode from 'vscode'
-import {
-  updateGlobalState,
-  getMessageIdSpan,
-  getMessageValueSpan,
-  getMessageHover,
-  isMessageReference,
-} from './global-state'
-import { getIdentifierRangeAtPosition } from './utils'
+import { updateGlobalState } from './global-state'
 import { updateDiagnosticCollection } from './diagnostic'
+import definitionProvider from './providers/definition'
+import hoverProvider from './providers/hover'
 
 const activate = (_context: vscode.ExtensionContext) => {
   vscode.workspace.textDocuments
@@ -40,53 +35,8 @@ const activate = (_context: vscode.ExtensionContext) => {
     updateDiagnosticCollection(event.document.uri.path)
   })
 
-  vscode.languages.registerDefinitionProvider('fluent', {
-    provideDefinition(document, position, _cancellationToken) {
-      const originSelectionRange = getIdentifierRangeAtPosition(document, position)
-      const messageIdentifier = document.getText(originSelectionRange)
-
-      if (isMessageReference(document.uri.path, messageIdentifier, document.offsetAt(position)) === false) {
-        return
-      }
-
-      const messageIdSpan = getMessageIdSpan(document.uri.path, messageIdentifier)
-      const messageIdPosition = document.positionAt(messageIdSpan.start)
-
-      const messageValueSpan = getMessageValueSpan(document.uri.path, messageIdentifier)
-      const messageValuePosition = document.positionAt(messageValueSpan.end)
-
-      return [
-        {
-          originSelectionRange,
-          targetUri: vscode.Uri.file(document.uri.path),
-          targetRange: new vscode.Range(
-            messageIdPosition,
-            messageValuePosition
-          ),
-          targetSelectionRange: new vscode.Range(
-            messageIdPosition,
-            new vscode.Position(messageIdPosition.line, messageIdentifier.length)
-          ),
-        },
-      ]
-    },
-  })
-
-  vscode.languages.registerHoverProvider('fluent', {
-    provideHover(document, position, _token) {
-      const messageIdentifier = document.getText(getIdentifierRangeAtPosition(document, position))
-
-      if (isMessageReference(document.uri.path, messageIdentifier, document.offsetAt(position)) === false) {
-        return
-      }
-
-      const content = getMessageHover(document.uri.path, messageIdentifier)
-
-      return {
-        contents: [content],
-      }
-    },
-  })
+  vscode.languages.registerDefinitionProvider('fluent', definitionProvider)
+  vscode.languages.registerHoverProvider('fluent', hoverProvider)
 }
 
 const deactivate = () => {
